baseimages: adapt nodejs flavour to debian baseimage

+++ baseimages/flavours/nodejs.sh
@@ -4,8 +4,9 @@ NODE_VERSION=8.11.2
 
 NVM_VERSION=0.33.11
 
-if [ -f "${CHROOT_DIR}/etc/alpine-release" ]; then
-    # Install dependencies for building the latest Node.js LTS.
+# Install dependencies for building the latest Node.js LTS.
+
+if is_alpine; then
     chroot_exec \
         apk --update add \
             curl \
@@ -15,15 +16,29 @@ if [ -f "${CHROOT_DIR}/etc/alpine-release" ]; then
             linux-headers \
             make \
             python
+elif is_debian; then
+    chroot_exec apt-get update
 
-    chroot_exec curl -LO "https://github.com/nodejs/node/archive/v${NODE_VERSION}.tar.gz"
+    chroot_exec \
+        apt-get install -y \
+            curl \
+            build-essential \
+            python
+else
+    fatal "unknown distro."
+fi
 
-    chroot_exec tar xzvf "v${NODE_VERSION}.tar.gz"
+# Download the latest Node.js LTS and build it.
 
-    chroot_exec sh -c "cd node-${NODE_VERSION} && ./configure --prefix=/usr && make -j${CORES_NUMBER} && make install"
+chroot_exec curl -LO "https://github.com/nodejs/node/archive/v${NODE_VERSION}.tar.gz"
 
-    # Cleanup
+chroot_exec tar xzvf "v${NODE_VERSION}.tar.gz"
 
+chroot_exec sh -c "cd node-${NODE_VERSION} && ./configure --prefix=/usr && make -j${CORES_NUMBER} && make install"
+
+# Cleanup
+
+if is_alpine; then
     chroot_exec \
         apk del \
             curl \
@@ -39,11 +54,23 @@ if [ -f "${CHROOT_DIR}/etc/alpine-release" ]; then
     chroot_exec \
         apk add libstdc++
 
-    rm -r "${CHROOT_DIR}/node-${NODE_VERSION}"
-
-    rm    "${CHROOT_DIR}/v${NODE_VERSION}.tar.gz"
-
     rm -f "${CHROOT_DIR}"/var/cache/apk/*
+elif is_debian; then
+    chroot_exec \
+        apt-get purge -y \
+            curl \
+            build-essential \
+            python
+
+    chroot_exec apt-get autoremove -y
+
+    chroot_exec apt-get clean
+
+    rm -rf "${CHROOT_DIR}"/var/lib/apt/lists/*
 else
-    info "doing nothing."
+    fatal "unknown distro."
 fi
+
+rm -r "${CHROOT_DIR}/node-${NODE_VERSION}"
+
+rm    "${CHROOT_DIR}/v${NODE_VERSION}.tar.gz"
